requirements.txt
flask==3.0.2
python-dotenv==1.0.1
pydantic==2.7.4

langchain==0.3.1
langchain-community==0.3.1
langchain-ollama==0.2.0

pymupdf==1.24.9
faiss-cpu==1.8.0.post1

orjson==3.10.7
tenacity==8.5.0
